migration of runtime configs from build to runtime key
