+++ truss/tests/test_config.py
@@ -2,14 +2,12 @@ import copy
 import tempfile
 from contextlib import nullcontext as does_not_raise
 from pathlib import Path
-from typing import Any, Dict
 
 import pytest
 import yaml
 
 from truss.base.custom_types import ModelFrameworkType
 from truss.base.trt_llm_config import (
-    TrussSpecDecMode,
     TrussTRTLLMQuantizationType,
 )
 from truss.base.truss_config import (
@@ -27,117 +25,6 @@ from truss.base.truss_config import (
     TrussConfig,
 )
 from truss.truss_handle.truss_handle import TrussHandle
-
-
-@pytest.fixture
-def default_config() -> Dict[str, Any]:
-    return {
-        "build_commands": [],
-        "environment_variables": {},
-        "external_package_dirs": [],
-        "model_metadata": {},
-        "model_name": None,
-        "python_version": "py39",
-        "requirements": [],
-        "resources": {
-            "accelerator": None,
-            "cpu": "1",
-            "memory": "2Gi",
-            "use_gpu": False,
-        },
-        "secrets": {},
-        "system_packages": [],
-    }
-
-
-@pytest.fixture
-def trtllm_config(default_config) -> Dict[str, Any]:
-    trtllm_config = default_config
-    trtllm_config["resources"] = {
-        "accelerator": Accelerator.L4.value,
-        "cpu": "1",
-        "memory": "24Gi",
-        "use_gpu": True,
-    }
-    trtllm_config["trt_llm"] = {
-        "build": {
-            "base_model": "llama",
-            "max_seq_len": 2048,
-            "max_batch_size": 512,
-            "checkpoint_repository": {
-                "source": "HF",
-                "repo": "meta/llama4-500B",
-            },
-            "gather_all_token_logits": False,
-        },
-        "runtime": {},
-    }
-    return trtllm_config
-
-
-@pytest.fixture
-def trtllm_spec_dec_config_full(trtllm_config) -> Dict[str, Any]:
-    spec_dec_config = copy.deepcopy(trtllm_config)
-    spec_dec_config["trt_llm"] = {
-        "build": {
-            "base_model": "llama",
-            "max_seq_len": 2048,
-            "max_batch_size": 512,
-            "checkpoint_repository": {
-                "source": "HF",
-                "repo": "meta/llama4-500B",
-            },
-            "plugin_configuration": {
-                "paged_kv_cache": True,
-                "gemm_plugin": "auto",
-                "use_paged_context_fmha": True,
-            },
-            "speculator": {
-                "speculative_decoding_mode": TrussSpecDecMode.DRAFT_EXTERNAL,
-                "num_draft_tokens": 4,
-                "build": {
-                    "base_model": "llama",
-                    "max_seq_len": 2048,
-                    "max_batch_size": 512,
-                    "checkpoint_repository": {
-                        "source": "HF",
-                        "repo": "meta/llama4-500B",
-                    },
-                },
-            },
-        },
-    }
-    return spec_dec_config
-
-
-@pytest.fixture
-def trtllm_spec_dec_config(trtllm_config) -> Dict[str, Any]:
-    spec_dec_config = copy.deepcopy(trtllm_config)
-    spec_dec_config["trt_llm"] = {
-        "build": {
-            "base_model": "llama",
-            "max_seq_len": 2048,
-            "max_batch_size": 512,
-            "checkpoint_repository": {
-                "source": "HF",
-                "repo": "meta/llama4-500B",
-            },
-            "plugin_configuration": {
-                "paged_kv_cache": True,
-                "gemm_plugin": "auto",
-                "use_paged_context_fmha": True,
-            },
-            "speculator": {
-                "speculative_decoding_mode": TrussSpecDecMode.DRAFT_EXTERNAL,
-                "num_draft_tokens": 4,
-                "checkpoint_repository": {
-                    "source": "HF",
-                    "repo": "meta/llama4-500B",
-                },
-            },
-        },
-    }
-    return spec_dec_config
 
 
 @pytest.mark.parametrize(
